1) Fixed a problem with the GC & threads: #define GC_THREADS 2) Added a separate thread for finalization. This avoids deadlocks & is recommended by the BoehmGC. Keep in mind all collection is disabled until the finalizers complete, which usually means collection occurs during the next pass. Most classes do not have finalizers, so collection delay is not a big issue. 3) Added the Thread to the ThreadGroup before spawning it. This keeps a non-native reference to the Thread, which is needed to avoid collection during a small startup window. Also, the main thread group is now statically referenced & Threads are removed from their ThreadGroup when they complete. 4) For Condition & Mutex, malloc/free are now used instead of XMLVM_ATOMIC_MALLOC/XMLVM_FREE
We still have a concern that the GC is collecting some instances too early.
Specifically, we had seen a Thread's Mutex finalized (without bringing it back to life), and then a method was called on it, so you'd see this error:
Error locking mutex! 22
Error unlocking mutex! 22

That error *may* be resolved with the addition of the static reference to the main ThreadGroup.

git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1899 0874c2aa-48f6-4a0f-8a6b-e32f4465a257