The garbage collector is fixed!
Native mutexes are now finalized after normal finalizers. The reason it is an exception is because mutexes are used for "synchronized" blocks, so in case a "synchronized" block occurs during finalization, the mutex should not have been destroyed yet.  In a normal JVM, mutexes do not share the same type of finalization as instances, which is why this case exists.

git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1903 0874c2aa-48f6-4a0f-8a6b-e32f4465a257